makefile: use pyproject.toml for installing pythemis
According to this [1] article, the correct command is

    pip install .

in the project's root. Let's try that. Also, the other option is

    python -m build --wheel

which builds the package, but doesn't install it. We can provide
something like `pythemis_build` for it for example.

Also, both commands require the `build` module to be installed.
Maybe we will need to add a new target `pythemis_install_deps`
or something that installs dependencies. But we will see.

[1]: https://godatadriven.com/blog/a-practical-guide-to-setuptools-and-pyproject-toml/